requirements.txt
streamlit>=1.37
pandas>=2.0
numpy>=1.25
pdfplumber==0.11.0
pillow>=10.0.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
python-levenshtein==0.25.1
pyyaml==6.0.2
